+++ src/cpp/solve/solve_LBFGS.cpp
@@ -14,31 +14,31 @@ struct LBFGSFunction
   const ProblemT &problem;
   explicit LBFGSFunction(const ProblemT &problem) : problem(problem) {}
 
-  float operator()(const Eigen::VectorXf &x, Eigen::VectorXf &grad) const
+  double operator()(const Eigen::VectorXd &x, Eigen::VectorXd &grad) const
   {
-    return static_cast<float>(problem.calc_score_and_grad(x, grad));
+    return static_cast<double>(problem.calc_score_and_grad(x, grad));
   }
 };
 
 template <typename ProblemT>
-void solve_LBFGS(const ProblemT &problem, std::vector<Eigen::VectorXf> &positions,
+void solve_LBFGS(const ProblemT &problem, std::vector<Eigen::VectorXd> &positions,
                  std::vector<std::pair<double, double>> &hist, Timer &timer,
                  const int MAX_ITER)
 {
   timer.start();
-  LBFGSpp::LBFGSParam<float> param;
+  LBFGSpp::LBFGSParam<double> param;
   param.m = 10;
   param.max_iterations = MAX_ITER;
   if (problem.modelType != ForceModelType::FR)
     param.max_iterations = 2 * MAX_ITER; // since non-FR models are harder to optimize, we allow more iterations
   param.epsilon = 1e-4;                  // to avoid line search failure
   param.epsilon_rel = 1e-4;              // to avoid line search failure
-  LBFGSpp::LBFGSSolver<float> solver(param);
+  LBFGSpp::LBFGSSolver<double> solver(param);
 
   LBFGSFunction<ProblemT> fun(problem);
   assert(!positions.empty());
-  Eigen::VectorXf x = positions.back();
-  float fx;
+  Eigen::VectorXd x = positions.back();
+  double fx;
 
   try
   {
